# containers/be-users-groups/src/main/resources/openapi.yaml
openapi: 3.0.1
info:
  title: Text API
  description: API for managing texts
  version: 1.0.0
servers:
  - url: http://0.0.0.0:8080/api/v1/text
paths:
  /:
    post:
      summary: Create a new text
      operationId: createText
      requestBody:
        description: Text to create
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Text'
      responses:
        '201':
          description: Created text object
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Text'
    get:
      summary: Get all texts
      operationId: getAllTexts
      responses:
        '200':
          description: A list of texts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Text'
  /{id}:
    get:
      summary: Get text by ID
      operationId: getTextById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: A text object
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Text'
        '404':
          description: Text not found
    put:
      summary: Update text by ID
      operationId: updateText
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        description: New text value
        required: true
        content:
          application/json:
            schema:
              type: string
      responses:
        '200':
          description: Updated text object
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Text'
        '404':
          description: Text not found
    delete:
      summary: Delete text by ID
      operationId: deleteText
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Text deleted successfully
        '404':
          description: Text not found
components:
  schemas:
    Text:
      type: object
      properties:
        id:
          type: integer
          format: int64
        content:
          type: string
      required:
        - content